Develop to main

+++ src/main/java/com/ryhma19/ilmastonmuutos/repository/V8Repository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ryhma19.ilmastonmuutos.data.V8;
 
-public interface V8Repository extends JpaRepository<V8, Integer> {
-    
+public interface V8Repository extends JpaRepository<V8, Long> {
+
 }
